delete piece.sortVerticesCCW

+++ include/reconstruction.h
@@ -52,12 +52,12 @@ public:
 	float boardHeight_;//20; 
 	float boardWidth_;//20;//note also the recommondation of static bodies (no more than 50!)
 
-	SfmlScreen* screen_; // valid also for silent mode for the last image
+	SfmlScreen* screen_; 
 	int screenHeight_;
 	int screenWidth_;
 //#pragma message("Please convert to double");
-	//float timeStep_ = 1 / 60; //1.0F / 60.0F;
-	double timeStep_ = 1.0 / 60.0F; //1.0F / 60.0F;
+	float timeStep_ = 1 / 60; //1.0F / 60.0F;
+	//double timeStep_ = 1.0 / 60.0F; //1.0F / 60.0F;
 	int velocityIterations_ = 6;
 	int positionIterations_ = 2;
